Aktualizace podzim 2022 – Java 17, Freemarker, aktualizace Springu

+++ src/main/java/cz/czechitas/java2webapps/lekce10/repository/TridaRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 /**
  */
 @Repository
-public interface TridaRepository  extends JpaRepository<Trida, Short> {
+public interface TridaRepository  extends JpaRepository<Trida, Integer> {
 }
